Add files via upload
The assignment presents a series of tasks designed to test and reinforce fundamental concepts in Python programming. Each task involves a specific coding challenge, and my goal is to fill in the blanks in the provided code to complete each task successfully. The tasks cover a range of topics, including variable initialization, mathematical operations, function creation and usage, list and dictionary manipulation, conditional statements, looping, and data analysis using pandas. To solve each task, I  need to replace the blank spaces in the code snippets with the appropriate values or code based on my understanding of Python programming concepts.